Add hostname to redirects for non-migrated pages w/o a trailing slash
Pages without a trailing slash are supposed to redirect to the page with a trailing slash. For example `page.18f.gov/lean-product-design` should redirect to `pages.18f.gov/lean-product-design`.

Unfortunatly, the redirect above was pointing to `pages-redirects.apps.cloud.gov/lean-product-design` since there was no hostname in the target for the rewrite. This commit adds a hostname to the target.